Move Joint class to separate file.
